+++ changelog.md
@@ -19,6 +19,8 @@
       - Fixed base monster delete/cut
       - Fixed default room/monster drop down not updating
       - Fixed Update all empty rooms to new default room type
+    - Virtual area editor:
+      - Fixed external exit editing for hidden flag
 
 ## v0.6.12 2018-12-03